Move vis.js GraphView adapter from core/graph to sdks/go
The VertexView/EdgeView/GraphView types and Graph.Render method exist
solely to project a graph into the vis-network JSON shape ({nodes,
edges} with integer IDs). They are not a graph-algorithm primitive
and have no consumer inside core/.

Move them to sdks/go, where Illuminate* already returns the
client.Graph callers want to render. Render is now a method on
*client.Graph with signature
  Render(func(string) int, func(*Vertex) string) GraphView

Drop core/graph/view.go and the generic Render from core/graph/model.go.
Move the test alongside, and remove the now-unused sort/strconv imports
from core/graph/graph_test.go.

Closes #267
